gradio-components/brickpointselector/demo/requirements.txt
gradio_brickpointselector